Integrate warrior skill tests
