+++ AGENTS.md
@@ -13,3 +13,5 @@ curl http://localhost:8000/debug.html | head -n 20
 이 과정을 통해 에러가 없는지 살펴본 뒤에 PR을 제출해 주세요.
 
 모듈화를 최우선적으로 고려하며 게임을 만들어주십시오.
+
+각 클래스의 새로운 스킬이 추가될 때마다, 각 클래스의 스킬을 통합적으로 테스트하는 곳에 새로 추가 후 테스트할 것.
